simplify client actions tests and fix typecheck

+++ packages/@collabkit/client/test/actions/deleteMessage.test.ts
@@ -45,7 +45,7 @@ test('deleteMessage', async () => {
   initComposer(store, { workspaceId, threadId, eventId: 'default' });
   store.workspaces[workspaceId].timeline[threadId] = {};
 
-  const { id } = await sync.sendMessage({
+  const { id: eventId } = await sync.sendMessage({
     appId,
     workspaceId,
     threadId,
@@ -59,19 +59,25 @@ test('deleteMessage', async () => {
     },
   });
 
-  await deleteMessage(store as Store, { workspaceId, threadId, eventId: id });
+  const id = await deleteMessage(store as Store, { workspaceId, threadId, eventId });
 
-  const events = await getTimeline({
+  expect(id).toBeDefined();
+  if (!id) throw new Error('ID is undefined');
+
+  const timeline = await getTimeline({
     appId,
     workspaceId,
     threadId,
   });
 
-  expect(events?.[1]).toStrictEqual({
+  expect(timeline).toBeDefined();
+  if (!timeline) throw new Error('Timeline is undefined');
+
+  expect(timeline[id]).toStrictEqual({
     id: expect.any(String),
     type: 'delete',
     body: '',
-    parentId: id,
+    parentId: eventId,
     createdAt: expect.any(Number),
     createdById: userId,
     mentions: [],
